Rename VirtualTable -> PigTable and move it back to piglet
